+++ lib/Extended_ast.mli
@@ -15,6 +15,8 @@ include module type of Parsetree
 
 type use_file = toplevel_phrase list
 
+type repl_file = repl_phrase list
+
 type 'a t =
   | Structure : structure t
   | Signature : signature t
@@ -22,12 +24,15 @@ type 'a t =
   | Core_type : core_type t
   | Module_type : module_type t
   | Expression : expression t
+  | Repl_file : repl_file t
 
 module Parse : sig
   val ast : 'a t -> Lexing.lexbuf -> 'a
 end
 
 val equal_core_type : core_type -> core_type -> bool
+
+val equal : 'a t -> 'a -> 'a -> bool
 
 val map : 'a t -> Ast_mapper.mapper -> 'a -> 'a